requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.8.4
Flask-Mail
Waitress
pycco